refactor: manage the view-related part in RacingGame class separately
